+++ ninecms/forms.py
@@ -6,14 +6,88 @@ __email__ = 'gkarak@9-dev.com'
 
 from django import forms
 from django.forms.models import inlineformset_factory
+from django.contrib.admin.widgets import FilteredSelectMultiple
 from django.contrib.auth.models import Group
 from django.utils.translation import ugettext_lazy as _
-from ninecms.models import Node, Image, File, Video
+from ninecms.models import Node, Image, File, Video, ContentBlock, PageType, TaxonomyTerm
 from ninecms.utils.sanitize import sanitize, ModelSanitizeForm
+
+
+class PageTypeForm(forms.ModelForm):
+    """ Override default page type form to show related blocks
+    https://www.lasolution.be/blog/related-manytomanyfield-django-admin-site.html
+    https://github.com/django/django/blob/master/django/contrib/admin/widgets.py#L24
+    """
+    # @todo make parent generic class for reverse related m2m fields (use a new meta prop)
+    blocks = forms.ModelMultipleChoiceField(
+        ContentBlock.objects.all(),
+        widget=FilteredSelectMultiple("Blocks", True),
+        required=False,
+    )
+
+    def __init__(self, *args, **kwargs):
+        """ Initialize form
+        :param args
+        :param kwargs
+        :return: None
+        """
+        super(PageTypeForm, self).__init__(*args, **kwargs)
+        if self.instance.pk:
+            self.initial['blocks'] = self.instance.blocks.values_list('pk', flat=True)
+        # from django.db.models import ManyToManyRel
+        # from django.contrib import admin
+        # rel = ManyToManyRel(ContentBlock, PageType)
+        # self.fields['blocks'].widget = RelatedFieldWidgetWrapper(self.fields['blocks'].widget, rel, admin.site)
+
+    def save(self, *args, **kwargs):
+        """ Handle saving of related blocks
+        :param args
+        :param kwargs
+        :return: instance
+        """
+        instance = super(PageTypeForm, self).save(*args, **kwargs)
+        if instance.pk:
+            for block in instance.blocks.all():
+                if block not in self.cleaned_data['blocks']:
+                    instance.blocks.remove(block)
+            for block in self.cleaned_data['blocks']:
+                if block not in instance.blocks.all():
+                    instance.blocks.add(block)
+        return instance
+
+    class Meta:
+        """ Meta class """
+        model = PageType
+        fields = ['name', 'description', 'guidelines', 'url_pattern']
+
+
+class ContentTypePermissionsForm(forms.Form):
+    """ Content type permissions form """
+    add_node = forms.ModelMultipleChoiceField(
+        queryset=Group.objects.all(),
+        widget=forms.CheckboxSelectMultiple,
+        required=False
+    )
+    change_node = forms.ModelMultipleChoiceField(
+        queryset=Group.objects.all(),
+        widget=forms.CheckboxSelectMultiple,
+        required=False
+    )
+    delete_node = forms.ModelMultipleChoiceField(
+        queryset=Group.objects.all(),
+        widget=forms.CheckboxSelectMultiple,
+        required=False
+    )
 
 
 class ContentNodeEditForm(forms.ModelForm):
     """ Node edit or create form """
+    terms = forms.ModelMultipleChoiceField(
+        TaxonomyTerm.objects.all(),
+        widget=FilteredSelectMultiple("Terms", True),
+        required=False,
+    )
+
     def __init__(self, *args, **kwargs):
         """ Get user object to check if has full_html permission
         Checks if current_user is already set (eg from ModelAdmin)
@@ -26,6 +100,8 @@ class ContentNodeEditForm(forms.ModelForm):
         except AttributeError:
             self.current_user = kwargs.pop('user', None)
         super(ContentNodeEditForm, self).__init__(*args, **kwargs)
+        if self.instance.pk:
+            self.initial['terms'] = self.instance.terms.values_list('pk', flat=True)
 
     def clean(self):
         """ Override clean form to bleach
@@ -40,6 +116,22 @@ class ContentNodeEditForm(forms.ModelForm):
             if field in cleaned_data:
                 cleaned_data[field] = sanitize(cleaned_data[field], full_html=full_html)
         return cleaned_data
+
+    def save(self, *args, **kwargs):
+        """ Handle saving of related terms
+        :param args
+        :param kwargs
+        :return: instance
+        """
+        instance = super(ContentNodeEditForm, self).save(*args, **kwargs)
+        if instance.pk:
+            for term in instance.terms.all():
+                if term not in self.cleaned_data['terms']:
+                    instance.terms.remove(term)
+            for term in self.cleaned_data['terms']:
+                if term not in instance.terms.all():
+                    instance.terms.add(term)
+        return instance
 
     class Meta:
         """ Form model and fields """
@@ -135,16 +227,3 @@ class SearchForm(forms.Form):
         if 'q' in cleaned_data:
             cleaned_data['q'] = sanitize(cleaned_data['q'], allow_html=False)
         return cleaned_data
-
-
-class ContentTypePermissionsForm(forms.Form):
-    """ Content type permissions form """
-    add_node = forms.ModelMultipleChoiceField(queryset=Group.objects.all(),
-                                              widget=forms.CheckboxSelectMultiple,
-                                              required=False)
-    change_node = forms.ModelMultipleChoiceField(queryset=Group.objects.all(),
-                                                 widget=forms.CheckboxSelectMultiple,
-                                                 required=False)
-    delete_node = forms.ModelMultipleChoiceField(queryset=Group.objects.all(),
-                                                 widget=forms.CheckboxSelectMultiple,
-                                                 required=False)